requirements.txt
Package            Version
------------------ ----------
certifi            2022.12.7
charset-normalizer 2.1.1
contourpy          1.1.0
cycler             0.11.0
filelock           3.9.0
fonttools          4.40.0
idna               3.4
Jinja2             3.1.2
kiwisolver         1.4.4
MarkupSafe         2.1.2
matplotlib         3.7.1
mpmath             1.2.1
networkx           3.0
numpy              1.25.0
opencv-python      4.7.0.72
packaging          23.1
Pillow             9.5.0
pip                23.1.2
pyparsing          3.1.0
python-dateutil    2.8.2
PyYAML             6.0
requests           2.28.1
scipy              1.10.1
setuptools         65.5.0
six                1.16.0
sympy              1.11.1
torch              2.0.1+cpu
torchaudio         2.0.2+cpu
torchvision        0.15.2+cpu
tqdm               4.65.0
typing_extensions  4.4.0
urllib3            1.26.13
